requirements.txt
discord.py
yt-dlp
wikipedia
aiohttp
openai
pymongo
python-dotenv